+++ docs/maintaining/tagging_examples/docker_runner.py
@@ -1,3 +1,5 @@
+# Copyright (c) Jupyter Development Team.
+# Distributed under the terms of the Modified BSD License.
 from tagging.utils.docker_runner import DockerRunner
 
 with DockerRunner("ubuntu") as container: