+++ tests/image_specific_tests/base-notebook/test_pandoc.py
@@ -1,0 +1,17 @@
+# Copyright (c) Jupyter Development Team.
+# Distributed under the terms of the Modified BSD License.
+import logging
+
+from tests.utils.tracked_container import TrackedContainer
+
+LOGGER = logging.getLogger(__name__)
+
+
+def test_pandoc(container: TrackedContainer) -> None:
+    """Pandoc shall be able to convert MD to HTML."""
+    logs = container.run_and_wait(
+        timeout=10,
+        tty=True,
+        command=["bash", "-c", 'echo "**BOLD**" | pandoc'],
+    )
+    assert "<p><strong>BOLD</strong></p>" in logs